Move uuid hasing to namui-type
