+++ namui-type/Cargo.toml
@@ -10,3 +10,5 @@ serde = { version = "1.0", features = ["derive", "rc"] }
 num = { version = "0.4.0", features = ["std"] }
 ordered-float = "2.10"
 auto_ops = "0.3"
+uuid = { version = "1.1", features = ["v4", "js", "macro-diagnostics", "serde"]}
+siphasher = "0.3"
