+++ namui-type/README.md
@@ -1,0 +1,1 @@
+The types, which even server can use.
