+++ namui/Cargo.toml
@@ -36,8 +36,6 @@ reqwest = "0.11"
 fluvio-wasm-timer = "0.2"
 textwrap = "0.15"
 unicode-segmentation = "1.9"
-uuid = { version = "1.1", features = ["v4", "js", "macro-diagnostics", "serde"]}
-siphasher = "0.3"
 namui-type = { path = "../namui-type" }
 
 [dev-dependencies]
